refactor(geo-data): extract pure core to geo-data-core.js
Workers can now import bundled data + helpers via relative ESM path.
geo-data.js public surface (window.GeoData, module.exports) is byte-compatible:
all 335 unit tests pass unchanged.
